doc(B-0883): design better git-crypt post-quantum lattice-based architecture and relocate TS files
Why:
- Backlog item B-0883 requires a design specification for Better Git-Crypt resolving access revocation, review visibility, and post-quantum readiness.
- Moving TS files out of tools/ and under src/Core.TypeScript/ aligns with the project's source-boundary hygiene conventions.

What:
- Wrote post-quantum lattice-based architecture design memo at `docs/research/2026-06-12-better-gitcrypt-post-quantum-lattice-based-architecture.md`.
- Closed backlog item `081KSNY2Z0008QG0R002JKH50A-better-gitcrypt-post-quantum-lattice-based-retraction-native.md`.
- Relocated `tools/bus/` to `src/Core.TypeScript/bus/`.
- Relocated `tools/backlog/generate-index.ts` to `src/Core.TypeScript/backlog/generate-index.ts`.
- Relocated `tools/substrate-claim-checker/` to `src/Core.TypeScript/substrate-claim-checker/`.
- Updated all TS file imports under `tools/` referencing `../bus` to point to `../../src/Core.TypeScript/bus`.
- Fixed `check-semantic-equivalence.test.ts` unused import and updated `inventory.test.ts` real repo scan check.
- Regenerated legacy backlog index `docs/BACKLOG.md`.

Proof:
- Verified that all 145 substrate claim checker tests pass.
- Verified that all 91 bus and claim tests pass.
- Verified C# and F# code builds successfully (0 warnings, 0 errors).
- Verified F# test suite (3,000+ tests pass).
- Verified typescript typecheck (`tsc --noEmit` has 0 errors).
- Verified `dotnet format --verify-no-changes` passes.

Agency-Signature-Version: 1
Agent: lior-gemini
Agent-Runtime: Gemini CLI
Agent-Model: Gemini
Credential-Identity: AceHack
Credential-Mode: shared
Human-Review: not-implied-by-credential
Human-Review-Evidence: none
Action-Mode: autonomous-fail-open
Task: B-0883
