+++ docs/BACKLOG.md
@@ -431,7 +431,7 @@ are closed (status: closed/done in frontmatter)._
 - [ ] **[B-0867.15](backlog/P1/081KSNY2Z0008QG0R002A785QR-15-per-host-adapters-github-gitlab-gitea-bitbucket-isomorphi.md)** Per-host adapters — GitHub + GitLab + Gitea/Codeberg/Forgejo + Bitbucket; isomorphic cross-host substrate per parent B-0867 pre-allocation
 - [x] **[B-0170.1](backlog/P1/081KSNY2Z0008QG0R002CK42QK-1-semantic-equivalence-drift-checker.md)** Substrate-claim-checker - semantic-equivalence-drift checker
 - [ ] **[B-0886.2](backlog/P1/081KSNY2Z0008QG0R002CR38D8-2-two-priority-axes-clarification-usb-not-2nd-on-operator-pe.md)** Two priority axes clarification — USB is NOT 2nd on operator-personal axis (in-front-of-eyes word-of-mouth + DevOps iteration); cloud is 2nd on that axis; both first-class per their respective axis
-- [ ] **[B-0883](backlog/P1/081KSNY2Z0008QG0R002JKH50A-better-gitcrypt-post-quantum-lattice-based-retraction-native.md)** Better git-crypt — post-quantum lattice-based (ML-KEM/Kyber/Saber/NTRU family) + retraction-native + diff-readable; supersedes 2026-04-21 git-crypt rejection; TS-first + Bouncy Castle pattern study
+- [x] **[B-0883](backlog/P1/081KSNY2Z0008QG0R002JKH50A-better-gitcrypt-post-quantum-lattice-based-retraction-native.md)** Better git-crypt — post-quantum lattice-based (ML-KEM/Kyber/Saber/NTRU family) + retraction-native + diff-readable; supersedes 2026-04-21 git-crypt rejection; TS-first + Bouncy Castle pattern study
 - [ ] **[B-0892](backlog/P1/081KSNY2Z0008QG0R002QA720J-three-lanes-concurrent-operating-discipline-encryption-plus-.md)** Three-lanes concurrent operating discipline — encryption + zflash + state-machine-substrate all moving forward until each lane's backlog drains; every idle tick is time not spent advancing a lane
 - [ ] **[B-0668.2](backlog/P1/081KSNY2Z0008QG0R002RPC7QZ-2-rx-dbsp-bridge-spec.md)** Compositional DBSP frame architecture: Rx-DBSP Bridge Specification
 - [ ] **[B-0885](backlog/P1/081KSNY2Z0008QG0R0030V5ZVS-agent-private-encrypted-state-otto-first-then-other-ais-asap.md)** Agent private encrypted state — give Otto private substrate first, then other AIs; ASAP per operator 2026-05-28
