+++ src/Core.TypeScript/substrate-claim-checker/check-counts.ts
