+++ src/Core.TypeScript/substrate-claim-checker/check-path-forms.test.ts
