+++ src/Core.TypeScript/substrate-claim-checker/check-path-forms.ts
