+++ src/Core.TypeScript/substrate-claim-checker/check-self-recursive.ts
