+++ src/Core.TypeScript/substrate-claim-checker/fixtures/_convention-drift-target-adr.md
