+++ src/Core.TypeScript/substrate-claim-checker/fixtures/convention-drift-no-reciprocal-marker.md
