+++ src/Core.TypeScript/substrate-claim-checker/fixtures/cross-surface-drift-9-vs-15.md
