+++ tools/github/poll-pr-gate-batch.ts
@@ -56,8 +56,8 @@
 import { spawn, spawnSync } from "node:child_process";
 import { fileURLToPath } from "node:url";
 import { dirname, resolve } from "node:path";
-import { allActiveClaims } from "../bus/claim.ts";
-import type { ClaimRecord } from "../bus/claim.ts";
+import { allActiveClaims } from "../../src/Core.TypeScript/bus/claim.ts";
+import type { ClaimRecord } from "../../src/Core.TypeScript/bus/claim.ts";
 
 export interface CheckCounts {
   ok: number;
